Update slick-joda-mapper to 2.6.0
To avoid implicit resolution conflicts, I have decided to mix-in a trait
for supplying instances for Joda in the Table classes.
I really want to mix in GenericJodaSupport provided by slick-joda-mapper.
But GenericJodaSupport is a class, so I can't mix-in.
I may modify slick-joda-mapper later. I think slick-joda-mapper's interface is not good.